+++ tests/test_setting_module.py
@@ -614,8 +614,8 @@ CASES: Cases = (
 
 @cases(CASES)
 def test(
-    input_: File | list[File],
+    files: File | list[File],
     expected: ExpectedIssue | list[ExpectedIssue] | None,
-    flake8_options,
+    options,
 ):
-    check_project(input_, expected, flake8_options)
+    check_project(files, expected, options)