+++ CHANGES.rst
@@ -5,6 +5,9 @@ Release notes
 0.1.0 (unreleased)
 ==================
 
+-   Switched from being a `flake8 <https://flake8.pycqa.org/en/latest/>`_
+    plugin to being a standalone tool, to support rules on non-Python files.
+
 -   Dropped Python 3.8 support, added Python 3.9+ support.
 
 -   Requires Scrapy 2.0.1+.